requirements.txt
# Core Dependencies
langchain==0.3.1
langchain-anthropic==0.3.0
anthropic==0.39.0
python-dotenv==1.0.1

# AST and Code Analysis
ast-grep-py==0.29.0
astroid==3.3.5
pylint==3.3.1
mypy==1.13.0
flake8==7.1.1
radon==6.0.1

# Git Integration
GitPython==3.1.43

# Vector Store and Embeddings
chromadb==0.5.20
sentence-transformers==3.3.1

# Security Scanning
bandit==1.8.0
safety==3.2.11

# Utilities
rich==13.9.4
pydantic==2.10.3
pyyaml==6.0.2

# UI (Optional)
streamlit==1.41.1
gradio==5.9.1

# Development
pytest==8.3.4
black==24.10.0
isort==5.13.2
